# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Pannaga API",
    "description": "Hyperlocal Monsoon Onset & Break Prediction System for India",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/advisories/{advisory_id}/translate/{lang}": {
      "get": {
        "tags": [
          "Advisories"
        ],
        "summary": "On-demand translation",
        "operationId": "translate_advisory_endpoint_api_v1_advisories__advisory_id__translate__lang__get",
        "parameters": [
          {
            "name": "advisory_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Advisory Id"
            }
          },
          {
            "name": "lang",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Lang"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/officer/queue": {
      "get": {
        "tags": [
          "Officer"
        ],
        "summary": "Pending advisory approval queue",
        "description": "Returns PENDING advisories within the officer's jurisdiction.\nCRITICAL advisories appear first.",
        "operationId": "approval_queue_api_v1_officer_queue_get",
        "parameters": [
          {
            "name": "block_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Block Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/officer/advisories/{advisory_id}/approve": {
      "post": {
        "tags": [
          "Officer"
        ],
        "summary": "Approve an advisory",
        "operationId": "approve_advisory_api_v1_officer_advisories__advisory_id__approve_post",
        "parameters": [
          {
            "name": "advisory_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Advisory Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApproveBody",
                "default": {}
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/officer/advisories/{advisory_id}/reject": {
      "post": {
        "tags": [
          "Officer"
        ],
        "summary": "Reject an advisory",
        "operationId": "reject_advisory_api_v1_officer_advisories__advisory_id__reject_post",
        "parameters": [
          {
            "name": "advisory_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Advisory Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApproveBody",
                "default": {}
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/officer/advisories/{advisory_id}/translate/{lang}": {
      "get": {
        "tags": [
          "Officer"
        ],
        "summary": "Translation preview",
        "operationId": "translation_preview_api_v1_officer_advisories__advisory_id__translate__lang__get",
        "parameters": [
          {
            "name": "advisory_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Advisory Id"
            }
          },
          {
            "name": "lang",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Lang"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/officer/broadcast": {
      "post": {
        "tags": [
          "Officer"
        ],
        "summary": "Broadcast approved advisories to farmers",
        "description": "Dispatch approved advisories to all farmers in the relevant panchayats.\n\nSafety:\n  - confirm=False \u2192 dry-run, returns what would be sent\n  - SIMULATED advisories \u2192 BLOCKED_SIMULATED, never reach provider\n  - Only officers with jurisdiction over the advisory's block may broadcast\n  - Enforces farmer-level 72h cooldown (CRITICAL alerts bypass)\n  - Enforces farmer consent / opt-out",
        "operationId": "broadcast_advisories_api_v1_officer_broadcast_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BroadcastBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApproveBody": {
        "properties": {
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "type": "object",
        "title": "ApproveBody"
      },
...
      "BroadcastBody": {
        "properties": {
          "advisory_ids": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Advisory Ids"
          },
          "channel": {
            "type": "string",
            "title": "Channel",
            "default": "SMS"
          },
          "confirm": {
            "type": "boolean",
            "title": "Confirm",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "advisory_ids"
        ],
        "title": "BroadcastBody"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}